fix: improve folder structure
