+++ src/Components/CreateQuiz/FormContainer/index.tsx
@@ -1,4 +1,3 @@
-import FormActions from "./FormActions";
 import QuizInfoForm from "./FormSteps/QuizInfoForm";
 import { Grid } from "@mui/material";
 import { useAtom } from "jotai";
